# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Olake is fastest open-source tool for replicating Databases to Apache Iceberg or Data Lakehouse.",
        "title": "Olake UI API",
        "contact": {
            "name": "OLake",
            "url": "https://olake.io/contact/",
            "email": "hello@olake.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
        "/api/v1/platform/releases": {
            "get": {
                "description": "Retrieve the latest platform release updates and metadata.",
                "tags": [
                    "Platform"
                ],
                "summary": "Get release updates",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit the number of releases returned",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ReleasesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    },
                    "500": {
                        "description": "failed to fetch release metadata",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/project/{projectid}/settings": {
            "get": {
                "description": "Retrieve the settings for a specific project.",
                "tags": [
                    "Project Settings"
                ],
                "summary": "Get project settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id (default is 123)",
                        "name": "projectid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project Settings fetched successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ProjectSettingsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "failed to validate request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400Response"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    },
                    "500": {
                        "description": "failed to retrieve project settings",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Create or update the settings for a project.",
                "tags": [
                    "Project Settings"
                ],
                "summary": "Update project settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id (default is 123)",
                        "name": "projectid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "project settings data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpsertProjectSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project Settings updated successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "failed to validate request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400Response"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    },
                    "413": {
                        "description": "payload too large",
                        "schema": {
                            "$ref": "#/definitions/dto.Error413Response"
                        }
                    },
                    "500": {
                        "description": "failed to update project settings",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.Error400Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Bad request"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.Error401Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Authentication required"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.Error413Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Payload too large"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.Error500Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Internal server error"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.JSONResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "operation completed successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.ProjectSettingsResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "project_id": {
                    "type": "string",
                    "example": "project-123"
                },
                "webhook_alert_url": {
                    "type": "string",
                    "example": "https://hooks.slack.com/services/xxx/yyy/zzz"
                }
            }
        },
        "dto.ReleaseMetadataResponse": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "dto.ReleaseTypeData": {
            "type": "object",
            "properties": {
                "current_version": {
                    "type": "string"
                },
                "releases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ReleaseMetadataResponse"
                    }
                }
            }
        },
        "dto.ReleasesResponse": {
            "type": "object",
            "properties": {
                "features": {
                    "$ref": "#/definitions/dto.ReleaseTypeData"
                },
                "olake": {
                    "$ref": "#/definitions/dto.ReleaseTypeData"
                },
                "olake_helm": {
                    "$ref": "#/definitions/dto.ReleaseTypeData"
                },
                "olake_ui_worker": {
                    "$ref": "#/definitions/dto.ReleaseTypeData"
                }
            }
        },
...
        "dto.UpsertProjectSettingsRequest": {
            "type": "object",
            "required": [
                "project_id"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "project_id": {
                    "type": "string",
                    "example": "project-123"
                },
                "webhook_alert_url": {
                    "type": "string",
                    "example": "https://hooks.slack.com/services/xxx/yyy/zzz"
                }
            }
        },
...
    },
...
}